requirements.txt
torch
transformers
datasets
accelerate
evaluate
rouge-score
sacrebleu
sentencepiece
pyarrow
pandas
numpy
scikit-learn
tqdm
pyyaml
requests